Text_To_Image_Generation/requirements.txt
streamlit
diffusers
torch
Pillow
transformers
accelerate
numpy